% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxDifferences.R
\name{fluxDifferences}
\alias{fluxDifferences}
\title{Report the fold change of fluxes between two models}
\usage{
fluxDifferences(model1, model2, fold2report = 2)
}
\author{
Daniel Camilo Osorio Hurtado <dcosorioh@unal.edu.co>
}

