% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDifferences.R
\name{plotDifferences}
\alias{plotDifferences}
\title{Plot the fold change of fluxes between two models into a bipartite graph.}
\usage{
plotDifferences(model1, model2, ...)
}
\arguments{
\item{model1}{A valid model for the \code{'sybil'} package.}

\item{model2}{A valid model for the \code{'sybil'} package. Must have the same reactions (reaction number and reaction identifiers) as \code{"model1"} with different restrictions.}
}
\description{
This functions calculates the fold change \code{"(fluxModel2/fluxModel1)-1"} for fluxes of two given metabolic models and plot it into a bipartite graph.
}
\author{
Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}

