% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2flux.R
\name{exp2flux}
\alias{exp2flux}
\title{Convert Gene Expression Data to FBA fluxes}
\usage{
exp2flux(model, expression, missing = "gmean", scale = FALSE)
}
\arguments{
\item{model}{A valid model for the \code{'sybil'} package}

\item{expression}{A valid ExpressionSet object (one by treatment)}

\item{missing}{A character string specifying the value to be used in missing cases; must be one of \code{'gmean'}, \code{'min'}, \code{'1q'}, \code{'mean'}, \code{'median'}, \code{'3q'}, or \code{'max'}}

\item{scale}{A boolean value to specify if data must be scaled}
}
\examples{
\dontrun{
# Required Libraries
library("sybil")
library("Biobase")

# Reference Data
data("Ec_core")
optimizeProb(Ec_core)

# Random Data
gE <- ExpressionSet(matrix(runif(50,1,100),dimnames = list(sample(Ec_core@allGenes,50),c())))

# Applying the function
modifiedModel <- exp2flux(model = Ec_core,expression = gE,missing = "gmean",scale = TRUE)

# Result
optimizeProb(modifiedModel)
}
}
\author{
Kelly Botero <kjboteroo@unal.edu.co> - 
Maintainer: Daniel Camilo Osorio Hurtado <dcosorioh@unal.edu.co>
}

