% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2flux.R
\name{exp2flux}
\alias{exp2flux}
\title{Convert gene expression data to FBA flux boundaries.}
\usage{
exp2flux(model, expression, outputID, organism = NULL, typeID = NULL,
  missing = "mean", scale = FALSE)
}
\arguments{
\item{model}{A valid model for the \code{'sybil'} package.}

\item{expression}{A valid ExpressionSet object (one by treatment).}

\item{organism}{A valid organism identifier for the KEGG database. List of valid organism identifiers are available in: http://rest.kegg.jp/list/organism.}

\item{typeID}{A string to define the type of ID used in GPR's. One of \code{"entrez"} or \code{"kegg"} must be given.}

\item{missing}{A character string specifying the value to be used in missing cases; must be one of \code{'min'}, \code{'1q'}, \code{'mean'}, \code{'median'}, \code{'3q'}, or \code{'max'}}

\item{scale}{A boolean value to specify if data must be scaled to assign a value of 1000 as max.}
}
\description{
Set the flux boundaries for each reaction based on the expression values of genes associated in their GPR rules.
}
\details{
This function computes the flux boundaries for each reaction based in their associated GPR.
The value is obtained as follows: When two or more genes are associated by AND operators in the GPR rule, the minima value of their associated expression values is selected.
The AND operator indicates that all asociated genes are required to the enzymatic complex formation. In that case, the gene with the smaller expression value acts as the enzyme complex formation limitant.
In turn, when two or more genes are associated by OR operators in the GPR rule, the sum of all their associated expression values is computed. The OR operator indicates that each gene can code an entire isozyme to act as reaction catalyst.
To missing gene expression values, the function assigns one of: \code{'min'}, \code{'1q'}, \code{'mean'}, \code{'median'}, \code{'3q'}, or \code{'max'} expression value computed from the genes associated to the same metabolic pathway.
In case of not possible pathway assignment to a gene, the value is calculated from all gene expression values.
The flux boundaries of exchange reactions are not modified.
}
\examples{
\dontrun{
# Loading a model
library("sybil")
library("Biobase")

# Original model:
data("Ec_core")

# Original model evaluation:
optimizeProb(Ec_core)

# Generating simulated expressionSets
expressionData <- matrix(data = runif(3*length(Ec_core@allGenes),min = 1,max = 100),
                         nrow = length(Ec_core@allGenes),
                         dimnames = list(c(Ec_core@allGenes),c()))
expressionData <- ExpressionSet(assayData = expressionData)

# Applying exp2flux
Ec_coreGE <- exp2flux(model = Ec_core,
                      expression = expressionData,
                      missing = "mean")
# Evaluating exp2flux model
optimizeProb(Ec_coreGE)
}
}
\author{
Daniel Camilo Osorio <dcosorioh@unal.edu.co> and Kelly Botero <kjboteroo@unal.edu.co>
}
