% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2flux.R
\name{exp2flux}
\alias{exp2flux}
\title{Convert Gene Expression Data to FBA fluxes}
\usage{
exp2flux(model, expression, organism = NULL, typeID = NULL,
  missing = "mean", scale = TRUE)
}
\arguments{
\item{model}{A valid model for the \code{'sybil'} package}

\item{expression}{A valid ExpressionSet object (one by treatment)}

\item{organism}{}

\item{typeID}{}

\item{missing}{A character string specifying the value to be used in missing cases; must be one of \code{'min'}, \code{'1q'}, \code{'mean'}, \code{'median'}, \code{'3q'}, or \code{'max'}}

\item{scale}{A boolean value to specify if data must be scaled}
}
\author{
Kelly Botero <kjboteroo@unal.edu.co> - Maintainer: Daniel Camilo Osorio <dcosorioh@unal.edu.co>
}

